Safari-611.1.10/LayoutTests/fast/css/MediaList-parameters.html

<!DOCTYPE html>
<html>
<head>
<style>
@media screen { a { text-weight: bold } }
</style>
</head>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that the parameters to MediaList.item() / deleteMedium() / appendMedium() are mandatory.");

var mediaList = document.getElementsByTagName("style")[0].sheet.media;
shouldBe("mediaList.__proto__", "MediaList.prototype");

shouldThrow("mediaList.item()", "'TypeError: Not enough arguments'");
shouldThrow("mediaList.deleteMedium()", "'TypeError: Not enough arguments'");
shouldThrow("mediaList.appendMedium()", "'TypeError: Not enough arguments'");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
