Safari-609.1.11.1/LayoutTests/fast/css/MediaList-parameters-expected.txt

Tests that the parameters to MediaList.item() / deleteMedium() / appendMedium() are mandatory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaList.__proto__ is MediaList.prototype
PASS mediaList.item() threw exception TypeError: Not enough arguments.
PASS mediaList.deleteMedium() threw exception TypeError: Not enough arguments.
PASS mediaList.appendMedium() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
